pyproject.toml
[project]
name = "reasoning-agents-framework"
version = "1.0.0"
description = "A modular framework for building multi-agent reasoning systems with A2A protocol and MCP tools"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["agents", "a2a", "mcp", "reasoning", "multi-agent", "langchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# =============================================================================
# Dependencies
# =============================================================================
dependencies = [
    # Agent-to-Agent Protocol
    "python-a2a>=0.1.0",
    
    # MCP (Model Context Protocol) Tools
    "fastmcp>=0.1.0",
    
    # LLM Integration
    "langchain-openai>=0.1.0",
    "langchain-mcp-adapters>=0.1.0",
    "langgraph>=0.1.0",
    
    # Database
    "duckdb>=0.9.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    
    # Configuration & Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "aiohttp>=3.9.0",
]

# =============================================================================
# Optional Dependencies
# =============================================================================
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

# =============================================================================
# Scripts (CLI entry points)
# =============================================================================
[project.scripts]
reasoning-agents = "core.cli:main"

# =============================================================================
# Build System
# =============================================================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "tools"]

# =============================================================================
# Tool Configurations
# =============================================================================
[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
